feat(suite-native): handle device pin setup in onboarding
